# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CutmanAI API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface AuthCredentials {
  email: string;
  /** @minLength 6 */
  password: string;
}

export interface UserResponse {
  id: number;
  email: string;
}

export type ReportSummaryStatus =
  (typeof ReportSummaryStatus)[keyof typeof ReportSummaryStatus];

export const ReportSummaryStatus = {
  pending: "pending",
  processing: "processing",
  complete: "complete",
  error: "error",
} as const;

export interface ReportSummary {
  id: number;
  fighterName?: string | null;
  videoSource?: string | null;
  status: ReportSummaryStatus;
  createdAt: string;
  errorMessage?: string | null;
}

export type ReportFullStatus =
  (typeof ReportFullStatus)[keyof typeof ReportFullStatus];

export const ReportFullStatus = {
  pending: "pending",
  processing: "processing",
  complete: "complete",
  error: "error",
} as const;

export interface ReportFull {
  id: number;
  fighterName?: string | null;
  videoSource?: string | null;
  status: ReportFullStatus;
  createdAt: string;
  reportContent?: string | null;
  errorMessage?: string | null;
}

export type ReportStatusStatus =
  (typeof ReportStatusStatus)[keyof typeof ReportStatusStatus];

export const ReportStatusStatus = {
  pending: "pending",
  processing: "processing",
  complete: "complete",
  error: "error",
} as const;

export interface ReportStatus {
  id: number;
  status: ReportStatusStatus;
  errorMessage?: string | null;
}

export type UploadVideoBodyOne = {
  video?: Blob;
  youtubeUrl?: string;
  fighterName?: string;
};

export type UploadVideoBodyTwo = {
  youtubeUrl?: string;
  fighterName?: string;
};

export type UpdateReportNameBody = {
  fighterName: string;
};