# lib/api-zod/src/generated/types/reportFull.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CutmanAI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ReportFullStatus } from "./reportFullStatus";

export interface ReportFull {
  id: number;
  fighterName?: string | null;
  videoSource?: string | null;
  status: ReportFullStatus;
  createdAt: Date;
  reportContent?: string | null;
  errorMessage?: string | null;
}