# lib/api-zod/src/generated/types/reportStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CutmanAI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ReportStatusStatus } from "./reportStatusStatus";

export interface ReportStatus {
  id: number;
  status: ReportStatusStatus;
  errorMessage?: string | null;
}