# lib/api-zod/src/generated/types/reportSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CutmanAI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ReportSummaryStatus } from "./reportSummaryStatus";

export interface ReportSummary {
  id: number;
  fighterName?: string | null;
  videoSource?: string | null;
  status: ReportSummaryStatus;
  createdAt: Date;
  errorMessage?: string | null;
}